Split section data and section methods
